Cleanup & attempt to fix loading orientation from json
