+++ .vscode/settings.json
@@ -36,6 +36,7 @@
     "plotly",
     "proxyurl",
     "remotedev",
+    "repr",
     "resp",
     "reversescale",
     "scatterplot",
@@ -48,33 +49,5 @@
     "ylabel",
     "yprop"
   ],
-  "importSorter.generalConfiguration.sortOnBeforeSave": true,
-  "importSorter.importStringConfiguration.maximumNumberOfImportExpressionsPerLine.type": "newLineEachExpressionAfterCountLimit",
-  "importSorter.importStringConfiguration.tabSize": 2,
-  "importSorter.importStringConfiguration.trailingComma": "multiLine",
-  "importSorter.sortConfiguration.removeUnusedImports": true,
-  "importSorter.sortConfiguration.customOrderingRules.rules": [
-    {
-      "regex": "^react(-dom)?$",
-      "orderLevel": 1,
-    },
-    {
-      "type": "importMember",
-      "regex": "^$",
-      "orderLevel": 5,
-      "disableSort": true
-    },
-    {
-      "regex": "^[^.@]",
-      "orderLevel": 10
-    },
-    {
-      "regex": "^[@]",
-      "orderLevel": 15
-    },
-    {
-      "regex": "^[.]",
-      "orderLevel": 30
-    }
-  ],
+  "editor.tabSize": 2,
 }
