+++ src/components/nglViewer/constants.ts
@@ -1,0 +1,93 @@
+export const BACKGROUND_COLOR = {
+  black: 'black',
+  white: 'white',
+};
+
+export const NGL_PARAMS = {
+  impostor: 'impostor',
+  quality: 'quality',
+  sampleLevel: 'sampleLevel',
+  theme: 'theme',
+  overview: 'overview',
+  rotateSpeed: 'rotateSpeed',
+  zoomSpeed: 'zoomSpeed',
+  panSpeed: 'panSpeed',
+  cameraFov: 'cameraFov',
+  cameraType: 'cameraType',
+  lightColor: 'lightColor',
+  lightIntensity: 'lightIntensity',
+  ambientColor: 'ambientColor',
+  ambientIntensity: 'ambientIntensity',
+  hoverTimeout: 'hoverTimeout',
+  backgroundColor: 'backgroundColor',
+  clipNear: 'clipNear',
+  clipFar: 'clipFar',
+  clipDist: 'clipDist',
+  fogNear: 'fogNear',
+  fogFar: 'fogFar',
+};
+
+// warning dont use underscore _ in constants!!!!!!
+export const OBJECT_TYPE = {
+  SPHERE: 'SPHERE',
+  COMPLEX: 'COMPLEX',
+  SURFACE: 'SURFACE',
+  DENSITY: 'DENSITY',
+  MOLECULE: 'MOLECULE',
+  PROTEIN: 'PROTEIN',
+  ARROW: 'ARROW',
+  CYLINDER: 'CYLINDER',
+  LINE: 'LINE',
+  EVENTMAP: 'EVENTMAP',
+  HOTSPOT: 'HOTSPOT',
+  MOLECULE_GROUP: 'MOLECULE-GROUP',
+};
+
+export const SELECTION_TYPE = {
+  PROTEIN: 'PROTEIN',
+  COMPLEX: 'COMPLEX',
+  SURFACE: 'SURFACE',
+  DENSITY: 'DENSITY',
+  LIGAND: 'LIGAND',
+  VECTOR: 'VECTOR',
+};
+
+export const MOL_REPRESENTATION = {
+  axes: 'axes',
+  backbone: 'backbone',
+  ballPlusStick: 'ball+stick',
+  base: 'base',
+  cartoon: 'cartoon',
+  contact: 'contact',
+  distance: 'distance',
+  helixorient: 'helixorient',
+  hyperball: 'hyperball',
+  label: 'label',
+  licorice: 'licorice',
+  line: 'line',
+  point: 'point',
+  ribbon: 'ribbon',
+  rocket: 'rocket',
+  rope: 'rope',
+  spacefill: 'spacefill',
+  surface: 'surface',
+  trace: 'trace',
+  tube: 'tube',
+  unitcell: 'unitcell',
+  validation: 'validation',
+};
+
+export const VIEWS = {
+  MAJOR_VIEW: 'major_view',
+  SUMMARY_VIEW: 'summary_view',
+  PANDDA_MAJOR: 'pandda_major',
+  PANDDA_SUMMARY: 'pandda_summary',
+};
+
+const prefix = 'REDUCERS_NGL_';
+
+export const ACTIONS = {
+  REMOVE_ALL_NGL_COMPONENTS: prefix + 'REMOVE_ALL_NGL_COMPONENTS',
+};
+
+export const MOL_REPRESENTATION_BUFFER = 'buffer';
